Allow quitting a game while paused
Pressing START with SELECT pressed while the game is paused will return
to the game mode select screen.

Since the "paused" strip is now too tall to look nice with a dark grey
background, change it to use a black background with lines at the top
and bottom so it still looks like it's on top of everything (objects
should not cover the text).