Save the new game state to set during a fade
Before this, if the new game state didn't come directly after the
current one (since it was incremented during a fade), it would have to
be set to 1 less than the new game state.

Because it would be setting the game state before a fade, disabling
objects mid-frame for the "paused" strip, which checked the current game
state, would not occur during a fade.