Allow the use of power-ups with the B button
Power-ups will change how a certain part of the game works (e.g. cut
cookie speeds in half) for a period of time. The currently selected
power-up will be used and removed when the player presses the B button.

The currently in-use power-up is shown in the middle of the status bar.