Prevent slowdown from too many actors
There were probably too many cookies anyway, so with this slightly lower
maximum cookie count, the game is still playable. My excuse will be that
it becomes a matter of endurance after this point.

Laser count cap is a little low for the double lasers power-up but
slowdown is worse than a missing laser here and there.