Save top 5 highest scores instead of just 1
If the new score is greater than or equal to any of the top 5 scores,
insert it into the list and shift all scores below it down (overwriting
the last).

Having 5 top scores means a single "high score" can't be shown on mode
select or game over screens since there are now 5. They should be shown
on the separate top scores screen.