Don't redraw things in the VBlank handler
Only redraw the score, number of cookies blasted, and hearts (lives)
when they change or are drawn over. The power-ups can change in a
variety of ways, so just draw them every frame, but not in the VBlank
handler. They can also be drawn at setup using the same subroutine.