Shoot "taller lasers" with fast lasers power-up
If the fast lasers power-up is in use, shoot a second laser above the
normal one to effectively make the lasers longer (still 2 lasers, each
able to take out a cookie).

This makes up for the lasers moving too fast to have a fair chance to
collide with a cookie horizontally.