Add action select screen and way to view top scores
The action select screen that appears after the title screen will give
the player the option to play the game or just to view top scores. From
there, the mode select screen appears for either action, and then either
the game starts or the appropriate top scores list is shown.